# apiDoc/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /purchases:
    post:
      tags:
      - Purchase
      summary: Purchase a dish from a restaurant
      operationId: purchaseDish
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseHistory'
  /transactions/top-users:
    get:
      tags:
      - Purchase
      summary: Find top x users by total transaction amount within a date range
      operationId: findTopTxUsers
      parameters:
      - name: top
        in: query
        description: "The specified top x (min = 1, max = 1000)"
        required: false
        schema:
          maximum: 1000
          minimum: 1
          type: integer
          format: int32
          default: 10
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTxAmount'
  /transactions/sum:
    get:
      tags:
      - Purchase
      summary: Find total number and dollar value of transactions that happened within
        a date range
      operationId: findTxNumbAmount
      parameters:
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TxNumbAmount'
  /transactions/max-restaurants:
    get:
      tags:
      - Purchase
      summary: "Find most popular restaurants by transaction volume, either by number\
        \ of transactions or transaction dollar value"
      operationId: findMaxTxRestaurants
      parameters:
      - name: byAmount
        in: query
        description: To indicate it's calculated by transaction amount or number of
          transactions
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantTxAmount'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /transactions/user-count:
    get:
      tags:
      - Purchase
      summary: Find total number of users who made transactions above or below $v
        within a date range
      operationId: getUserCount
      parameters:
      - name: amount
        in: query
        description: The specified transaction amount
        required: true
        schema:
          minimum: 0
          type: number
      - name: fromDate
        in: query
        description: "The specified start date (format is MM/dd/yyyy), eg: 12/25/2019"
        required: true
        schema:
          type: string
          format: date
      - name: toDate
        in: query
        description: "The specified end date (format is MM/dd/yyyy), eg: 02/14/2020"
        required: true
        schema:
          type: string
          format: date
      - name: lessThan
        in: query
        description: To indicate less or more than the specified transaction amount
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Count'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
  /open-hours:
    get:
      tags:
      - OpenHours
      summary: Find open hours by pagination
      operationId: findOpenHours
      parameters:
      - name: page
        in: query
        description: "The specified page, start from `0`"
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: The number of records in the page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
  /open-hours/findByTime:
    get:
      tags:
      - OpenHours
      summary: Find open hours by time
      operationId: findOpenHoursByTime
      parameters:
      - name: time
        in: query
        description: "The specified time (format is HH:mm), eg: 18:30"
        required: true
        schema:
          type: string
      - name: dayOfWeek
        in: query
        description: "A day of week (Sun = 0, Sat = 6)"
        required: false
        schema:
          maximum: 6
          minimum: 0
          type: integer
          format: int32
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
  /open-hours/findByRestaurant:
    get:
      tags:
      - OpenHours
      summary: Find open hours by restaurant ID
      operationId: findOpenHoursByRestaurant
      parameters:
      - name: restaurantId
        in: query
        description: The specified restaurant ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
...
components:
  schemas:
    Purchase:
      required:
      - menuId
      - restaurantId
      type: object
      properties:
        userId:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        menuId:
          type: integer
          format: int64
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        msg:
          type: string
    PurchaseHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        menuId:
          type: integer
          format: int64
        dishName:
          type: string
        transactionAmount:
          type: number
        transactionDate:
          type: string
          format: date-time
    UserTxAmount:
      type: object
      properties:
        userName:
          type: string
        userId:
          type: integer
          format: int64
        txAmount:
          type: number
    TxNumbAmount:
      type: object
      properties:
        txNumb:
          type: integer
          format: int64
        txAmount:
          type: number
    RestaurantTxAmount:
      type: object
      properties:
        restaurantName:
          type: string
        restaurantId:
          type: integer
          format: int64
        txNumb:
          type: integer
          format: int64
        txAmount:
          type: number
    Count:
      type: object
      properties:
        numb:
          type: integer
          format: int64
...
    Restaurant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cashBalance:
          type: number
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
    OpenHours:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurant:
          $ref: '#/components/schemas/Restaurant'
        dayOfWeek:
          type: integer
          format: int32
        openTime:
          $ref: '#/components/schemas/LocalTime'
        closedTime:
          $ref: '#/components/schemas/LocalTime'
        openPeriod:
          type: integer
          format: int32
...